Added some map and sets tips

+++ KattisPractices/wilson/cd.cpp
@@ -8,27 +8,27 @@ using namespace std;
 int main () {
     int numJack, numJill;
     while (true) {
-        unordered_set<long long> jack;
-        long long count = 0;
+        unordered_set<int> jack;
+        int count = 0;
         scanf("%d %d", &numJack, &numJill);
         if (numJack == 0 && numJill == 0) {
             break;
         }
         
-        for (long long i = 0; i < numJack; i++) {
-            long long num;
-            scanf("%lld", &num);
+        for (int i = 0; i < numJack; i++) {
+            int num;
+            scanf("%d", &num);
             jack.insert(num);
         }
         
-        for (long long i = 0; i < numJill; i++) {
-            long long num;
-            scanf("%lld", &num);
+        for (int i = 0; i < numJill; i++) {
+            int num;
+            scanf("%d", &num);
             if (jack.find(num) != jack.end()) {
                 count++;
             }
         }
-        printf("%lld\n", count);
+        printf("%d\n", count);
     }
 }
